fix(tests): correct tests for form validation
